refactor consultas.py
